+++ test-resources/script_with_overlapping_opts.clj
@@ -1,1 +1,1 @@
-(prn (first *command-line-args*))
+(prn *command-line-args*)